Add a canvas solution for #78
Instead of using turf to compute polygon union then intersection between
temperatures and aridity unified polygons (which DOES NOT WORK) we do
the same but with composition operation in canvas.

The key is to have XOR between aridity & temperatures areas. Then the
resulting drawing is the zone we want to hide when the user filtered
some features from the geojson. See components/organisms/Atlas/canvas.js
for more details.